Shortened participants of our 2012 Majug survey
The full names are not in the repo for privacy reasons.

Modus of declaring the winner: the last two characters of the commit
hash (hex) will be converted to dec and used to determine the winner
position by applying it modulo.

Note: the first line has position zero, so there is an offset of 1.

Formula to determine winner in Groovy
groovy -e 'println "the winner is in line " + ((0xAA % 42) + 1)'